Fixed C style typecasting
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   replacement/bip/bip.cc
	modified:   replacement/lfu/lfu.cc
	modified:   replacement/lip/lip.cc
	modified:   replacement/mru/mru.cc
	modified:   replacement/nmru/nmru.cc
	modified:   replacement/pseudoLRU/pseudoLRU.cc
	modified:   replacement/random/random.cc

+++ replacement/mru/mru.cc
@@ -13,7 +13,7 @@ void CACHE::initialize_replacement()
 {
 	// Initialize the MRU position
         MRU::mru_ways[this] = std::vector<uint32_t>();
-	for(uint32_t idx = 0; idx < NUM_SET; idx++)
+	for(size_t idx = 0; idx < NUM_SET; idx++)
 		MRU::mru_ways[this].push_back(0);
 }
